refactor: adopt monorepo structure
This adopts Lerna to assist with building a monorepo. I looked at pnpm
and Yarn workspaces too--while these do some monorepo things, both are
more primitive (and can actually be used by) Lerna, so I just stuck with
that.

As referenced in the issue, with this change the `protocol` gets
symlinked into the other packages locally. I removed the separate
generator package so that the `protocol` produces all the things it
needs without having an extra dependency. Then, when we want to publish,
one can `npm run version -- <version>`, which will run `lerna version`
to update the version of all packages and push their tags. Packages
get published from the Github workflow.

Fixes #260